apps/api/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pytest
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
requests
opencv-python-headless
numpy
httpx
pytesseract
Pillow
